Major speed improvement for .fit()
.fit() previously relied on __dictionary_extractor and __calculate_slopes_intercept() to assemble a dataset from the gathered_data dict and compute the least squares solution, repeating this process for each landmark or pixel value. However, this was very inefficient, requiring n repeats, where n is the number of points or pixels. 

A far faster solution is to take a multivariate approach and compute the least squares solution for all points or pixels simultaneously. Mathematically, this produces the exact same output as the original approach, but is orders of magnitudes faster. A new function, __assemble_multivariate, creates the multivariate arrays for points and pixel values, before passing to an improved version of __calculate_slopes_intercept. The returned arrays are reshaped to match the input dimensions and maintain compatibility with the .predict() method.